# cryptobyte: AddUint*LengthPrefixed API perfomance optimization with sync.Pool