--- a/cryptobyte/builder.go
+++ b/cryptobyte/builder.go
@@ -1,5 +1,5 @@
-// Use of this source code is governed by a BSD-style
 // Copyright 2017 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
 package cryptobyte